onclite: Restore better load balancing on boot
As it turned out, allowing background, system-background, restricted,
foreground and top-app to use all 8 cores results in severe jitter and lag
whenever background tasks are running, whether it be on boot, or during waking up
from sleep.

This commit readds pinning background to cores 0-1, system-background and
restricted to cores 0-3, foreground to cores 0-6 and top-app to cores 0-7.

Change-Id: Idb76879e75e05b437600f0b689c260c674eb7eea

Co-authored-by: Zile995 <stefan.zivkovic995@gmail.com>
Signed-off-by: GeoPD <geoemmanuelpd2001@gmail.com>
Signed-off-by: AlexThundrous <alexthundrous2104@gmail.com>
Signed-off-by: Risti766 <wahhap10@gmail.com>